+++ R/FAT2DIC_function.R
@@ -269,7 +269,7 @@ FAT2DIC=function(f1,
   print(anava1,na.print = "")
   cat("\n")
 
-  if(transf==1 && norm1$p.value<0.05 | transf==1 && indep$p.value<0.05 | transf==1 &&homog1$p.value<0.05){
+  if(transf==1 && norm1$p.value<0.05 | transf==1 && indep$p.value<0.05 | transf==1 && homog1$p.value<0.05){
     message("\nYour analysis is not valid, suggests using a non-parametric test and try to transform the data\n")}else{}
   if(transf != 1 && norm1$p.value<0.05 | transf!=1 && indep$p.value<0.05 | transf!=1 && homog1$p.value<0.05){
     message("\nYour analysis is not valid\n")}else{}